M_CHATBOTfinal/requirements.txt
python-dotenv
langchain
pinecone-client
sentence-transformers
PyPDF2
groq
langchain-community
langchain_groq
langchain_pinecone
sarvamai